course_project/api/clokwise/requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deptry==0.19.1
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-jsonform==2.22.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.0
loguru==0.7.2
phonenumbers==8.13.43
prompt_toolkit==3.0.47
PyJWT==2.9.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
win32-setctime==1.1.0
gunicorn==23.0.0
packaging==24.1
h11==0.14.0
uvicorn==0.30.6
pillow==10.4.0
psycopg2==2.9.9
fcm-django==2.2.1
celery==5.4.0
Redis==5.0.8 
channels
daphne
channels_redis
uwicorn
python-magic